requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
pdfplumber
Pillow
scikit-learn
joblib
python-magic    # optional, for mime detection
pandas
fitz
pdf2image
PyMuPDF 
reportlab
secure-smtplib